+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog][docs-changelog], and the version adher
 
 
 ## Unreleased
+
+## [0.5.0][changes-0.5.0] - 2025-12-17
 ### Added
 - Search replication members method to InstanceClient class.
 - Ability to update user attrs by the InstanceClient class.
@@ -67,5 +69,6 @@ The format is based on [Keep a Changelog][docs-changelog], and the version adher
 [changes-0.4.0]: https://github.com/canonical/mysql-shell-client/compare/0.3.0...0.4.0
 [changes-0.4.1]: https://github.com/canonical/mysql-shell-client/compare/0.4.0...0.4.1
 [changes-0.4.2]: https://github.com/canonical/mysql-shell-client/compare/0.4.1...0.4.2
+[changes-0.5.0]: https://github.com/canonical/mysql-shell-client/compare/0.4.2...0.5.0
 [docs-changelog]: https://keepachangelog.com/en/1.0.0/
 [docs-semver]: https://semver.org/spec/v2.0.0.html